+++ app/middleware/cors.js
@@ -1,8 +1,0 @@
-exports = module.exports = function() {
-  
-  return function(options) {
-    return require('cors')(options);
-  };
-};
-
-exports['@implements'] = 'http://i.bixbyjs.org/http/middleware/cors';