+++ CHANGELOG.md
@@ -13,6 +13,8 @@ to create store in non-development environment.
 ### Removed
 - Removed use of cookie-based session middleware, which was used when no session
 store was available, from `middleware/session` component.
+- Removed `middleware/cors` component, as it has no dependencies which require
+injection.
 
 ## [0.2.3] - 2021-11-17
 ### Added